+++ examples/flowchart/VerticalFlowChart.tsx
@@ -70,7 +70,7 @@ export const VerticalFlowChart = ({ fref, chartData }: MilestoneStory) => {
                         1. Beginning
                     </BoxedLabel>
                 </MilestoneMotion>
-                <MilestoneMotion initial={'hidden'} initialOn={'A2'}>
+                <MilestoneMotion initial={'hidden'} initialOn={'A'} transition={{ delay: 0.5 }}>
                     <Label position={[145, y1]} size={[200, 20]} className={'explanation'}>
                         Usually, we start at the beginning.
                     </Label>
@@ -82,27 +82,20 @@ export const VerticalFlowChart = ({ fref, chartData }: MilestoneStory) => {
                             [60, y2 - 19],
                         ]}
                         units={'absolute'}
-                    />
-                </MilestoneMotion>
-                <MilestoneMotion initial={'hidden'} initialOn={'B'} transition={{ delay: 0.45 }}>
-                    <FlowPath
-                        points={[
-                            [60, y1 + 14],
-                            [60, y2 - 19],
-                        ]}
-                        units={'absolute'}
                         markerEnd={'arrowV'}
                     />
+                </MilestoneMotion>
+                <MilestoneMotion initial={'hidden'} initialOn={'B'} transition={{ delay: 0.5 }}>
                     <BoxedLabel position={[60, y2]} size={[120, 28]} className={'B'}>
                         2. Middle
                     </BoxedLabel>
                 </MilestoneMotion>
-                <MilestoneMotion initial={'hidden'} initialOn={'B2'}>
+                <MilestoneMotion initial={'hidden'} initialOn={'B'} transition={{ delay: 1 }}>
                     <Label position={[145, y2 - 10]} size={[200, 20]} className={'explanation'}>
                         The middle stage is often the longest.
                     </Label>
                 </MilestoneMotion>
-                <MilestoneMotion initial={'hidden'} initialOn={'B3'}>
+                <MilestoneMotion initial={'hidden'} initialOn={'B'} transition={{ delay: 1.5 }}>
                     <Label position={[145, y2 + 10]} size={[200, 20]} className={'explanation'}>
                         It may contain many sub-stages.
                     </Label>
@@ -114,22 +107,15 @@ export const VerticalFlowChart = ({ fref, chartData }: MilestoneStory) => {
                             [60, y3 - 19],
                         ]}
                         units={'absolute'}
-                    />
-                </MilestoneMotion>
-                <MilestoneMotion initial={'hidden'} initialOn={'C'} transition={{ delay: 0.45 }}>
-                    <FlowPath
-                        points={[
-                            [60, y2 + 14],
-                            [60, y3 - 19],
-                        ]}
-                        units={'absolute'}
                         markerEnd={'arrowV'}
                     />
+                </MilestoneMotion>
+                <MilestoneMotion initial={'hidden'} initialOn={'C'} transition={{ delay: 0.5 }}>
                     <BoxedLabel position={[60, y3]} size={[120, 28]} className={'C'}>
                         3. End
                     </BoxedLabel>
                 </MilestoneMotion>
-                <MilestoneMotion initial={'hidden'} initialOn={'C2'}>
+                <MilestoneMotion initial={'hidden'} initialOn={'C'} transition={{ delay: 1 }}>
                     <Label position={[145, y3]} size={[200, 20]} className={'explanation'}>
                         Finally, we reach a conclusion.
                     </Label>