+++ packages/annotation/src/labels/index.tsx
@@ -1,4 +1,0 @@
-export * from './BoxedLabel'
-export * from './BoxedTitle'
-export * from './Paragraph'
-export * from './types'