+++ packages/annotation/src/lines/index.tsx
@@ -1,4 +1,0 @@
-export * from './BraceLabel'
-export * from './BracketLabel'
-export * from './LineLabel'
-export * from './types'