+++ packages/annotation/src/typography/BracketLabel.tsx
@@ -1,29 +1,32 @@
-import { LineLabelProps } from './types'
-import { Line, useDimensions, useScales, Typography, getClassName } from '@chsk/core'
-import { getLineAbsolutePositions } from './utils'
+import { useDimensions, useScales, Typography, Path, getClassName } from '@chsk/core'
+import { getLineAbsolutePositions, getBracketPositions } from './lines'
+import { BracketLabelProps } from './types'
 
-export const LineLabel = ({
+export const BracketLabel = ({
     // settings for line
+    variant = 'right',
     start,
     end,
     units = 'view',
     expansion = [0, 0],
+    tickSize = 5,
     lineStyle,
     // position of text
     translate = [0, -8],
     align = 0.5,
     angle = 0,
-    markerStart,
-    markerEnd,
     textStyle,
     //
     className,
     style,
     setRole = true,
     children,
-}: LineLabelProps) => {
+}: BracketLabelProps) => {
     const { size } = useDimensions()
     const scales = useScales()
+
+    // toggle left-handed and right-handed symbol via tickSize
+    tickSize = variant === 'left' ? -tickSize : tickSize
 
     const { lineStart, lineEnd } = getLineAbsolutePositions({
         start,
@@ -33,22 +36,26 @@ export const LineLabel = ({
         scales,
         size,
     })
+    const { tickStart, tickEnd } = getBracketPositions({
+        start: lineStart,
+        end: lineEnd,
+        size: tickSize,
+    })
     const textPos: [number, number] = [
         lineStart[0] + translate[0] + (lineEnd[0] - lineStart[0]) * align,
         lineStart[1] + translate[1] + (lineEnd[1] - lineStart[1]) * align,
     ]
-    const compositeClassName = getClassName('line-label', className)
+    const compositeClassName = getClassName('bracket-label', className)
 
     return (
-        <g style={style} className={className} role={setRole ? 'line-label' : undefined}>
-            <Line
-                variant={'line'}
-                x1={lineStart[0]}
-                y1={lineStart[1]}
-                x2={lineEnd[0]}
-                y2={lineEnd[1]}
-                markerStart={markerStart}
-                markerEnd={markerEnd}
+        <g
+            style={style}
+            className={compositeClassName}
+            role={setRole ? 'bracket-label-' + variant : undefined}
+        >
+            <Path
+                variant={'bracket'}
+                points={[tickStart, lineStart, lineEnd, tickEnd]}
                 className={compositeClassName}
                 style={lineStyle}
                 setRole={setRole}