+++ packages/annotation/src/typography/types.ts
@@ -10,8 +10,14 @@ import {
     TranslateSpec,
     TypographyProps,
     AngleUnit,
+    SvgElementProps,
+    LineProps,
+    PositionUnits,
+    TwoSideSizeSpec,
 } from '@chsk/core'
 import { ReactNode } from 'react'
+
+/** labels with boxes */
 
 interface BoxedTextProps {
     /** expansion of box size */
@@ -54,6 +60,49 @@ export interface BoxedTitleProps extends SvgElementVariantProps, BoxedTextProps 
     offset?: number
 }
 
+/** labels with lines */
+
+export interface LineLabelProps
+    extends SvgElementProps,
+        Pick<LineProps, 'markerStart' | 'markerEnd'> {
+    /** start position */
+    start: PositionSpec
+    /** end position */
+    end: PositionSpec
+    /** units for start and end positions */
+    units?: PositionUnits
+    /** expansion of interval (multiples of bandwidth) */
+    expansion?: TwoSideSizeSpec
+    /** alignment for text label in [0, 1] */
+    align?: number
+    /** translation for text label */
+    translate?: TranslateSpec
+    /** rotation of text label */
+    angle?: number
+    /** padding for text label */
+    padding?: SizeSpec
+    /** styles for line */
+    lineStyle?: CssProps
+    /** styles for text */
+    textStyle?: CssProps
+    /** children components */
+    children?: ReactNode
+}
+
+export interface BracketLabelProps extends Omit<LineLabelProps, 'markerStart' | 'markerEnd'> {
+    /** left- or right- handed ticks */
+    variant?: 'right' | 'left'
+    /** size of marker at the end of line */
+    tickSize?: number
+}
+
+export interface BraceLabelProps extends BracketLabelProps {
+    /** smoothness of brace edges */
+    braceR?: number
+}
+
+/** Miscellaneous text */
+
 export interface ParagraphProps extends Omit<TypographyProps, 'variant'> {
     /** size of individual lines (width, height) */
     size: SizeSpec
@@ -65,4 +114,11 @@ export interface ParagraphProps extends Omit<TypographyProps, 'variant'> {
     letterBaseWidths?: 'serif' | 'sans'
     /** letter widths */
     letterWidths?: Record<string, number>
+}
+
+export interface FlowTypographyProps extends TypographyProps {
+    /** duration of animation (seconds) */
+    duration?: number
+    /** animation rate (seconds per character) */
+    rate?: number
 }