+++ packages/annotation/tests/typography/BracketLabel.test.tsx
@@ -1,66 +1,60 @@
 import { render, screen } from '@testing-library/react'
 import { Chart, View } from '@chsk/core'
-import { LineLabel } from '../../src/lines'
+import { BracketLabel } from '../../src/'
 import { chartProps, viewProps } from '../props'
 
-describe('LineLabel', () => {
+describe('BracketLabel', () => {
     it('creates a line with absolute coordinates', () => {
         render(
             <Chart {...chartProps}>
                 <View {...viewProps}>
-                    <LineLabel start={[0, 0]} end={[60, 0]} units={'absolute'}>
+                    <BracketLabel start={[0, 0]} end={[60, 0]} units={'absolute'}>
                         Label
-                    </LineLabel>
+                    </BracketLabel>
                 </View>
             </Chart>
         )
-        const result = screen.getByRole('line-label')
-        const line = result.querySelector('line')
-        // the line should go from (0, 0) to (60, 0)
-        expect(line?.getAttribute('x1')).toBe('0')
-        expect(line?.getAttribute('x2')).toBe('60')
-        expect(line?.getAttribute('y2')).toBe('0')
+        const result = screen.getByRole('bracket')
+        expect(result?.getAttribute('class')).toContain('bracket')
     })
 
     it('creates a line with relative coordinates', () => {
         render(
             <Chart {...chartProps}>
                 <View {...viewProps}>
-                    <LineLabel start={[0, 0]} end={[0.5, 0]} units={'relative'}>
+                    <BracketLabel start={[0, 0]} end={[0.5, 0]} units={'relative'}>
                         Label
-                    </LineLabel>
+                    </BracketLabel>
                 </View>
             </Chart>
         )
-        const result = screen.getByRole('line')
-        expect(Number(result.getAttribute('x2'))).toBeGreaterThan(100)
+        expect(screen.getByRole('bracket')).toBeDefined()
     })
 
     it('creates a line with view coordinates', () => {
         render(
             <Chart {...chartProps}>
                 <View {...viewProps}>
-                    <LineLabel start={[0, 0]} end={[80, 0]} units={'relative'}>
+                    <BracketLabel start={[0, 0]} end={[80, 0]} units={'view'}>
                         Label
-                    </LineLabel>
+                    </BracketLabel>
                 </View>
             </Chart>
         )
-        const result = screen.getByRole('line')
-        expect(Number(result.getAttribute('x2'))).toBeGreaterThan(100)
+        expect(screen.getAllByRole('bracket')).toBeDefined()
     })
 
     it('creates a line with text rotation', () => {
         render(
             <Chart {...chartProps}>
                 <View {...viewProps}>
-                    <LineLabel start={[0, 0]} end={[1, 0]} angle={45}>
+                    <BracketLabel start={[0, 0]} end={[1, 0]} angle={45}>
                         Label
-                    </LineLabel>
+                    </BracketLabel>
                 </View>
             </Chart>
         )
-        const result = screen.getByRole('line-label')
+        const result = screen.getByRole('bracket-label-right')
         expect(result.querySelector('text')?.closest('g')?.getAttribute('style')).toContain(
             'rotate(45'
         )
@@ -70,13 +64,13 @@ describe('LineLabel', () => {
         render(
             <Chart {...chartProps}>
                 <View {...viewProps}>
-                    <LineLabel start={[0, 0]} end={[1, 0]} setRole={false}>
+                    <BracketLabel start={[0, 0]} end={[1, 0]} setRole={false}>
                         Label
-                    </LineLabel>
+                    </BracketLabel>
                 </View>
             </Chart>
         )
-        const result = screen.queryByRole('line')
+        const result = screen.queryByRole('bracket')
         expect(result).toBeNull()
     })
 
@@ -84,16 +78,16 @@ describe('LineLabel', () => {
         render(
             <Chart {...chartProps}>
                 <View {...viewProps}>
-                    <LineLabel start={[0, 0]} end={[60, 0]} className={'custom'}>
-                        Label
-                    </LineLabel>
+                    <BracketLabel start={[0, 0]} end={[60, 0]} className={'custom'}>
+                        abc
+                    </BracketLabel>
                 </View>
             </Chart>
         )
-        const result = screen.getByRole('line-label')
+        const result = screen.getByRole('bracket-label-right')
         const text = result.querySelector('text')
-        const line = result.querySelector('line')
-        expect(text?.getAttribute('class')).toContain('label lineLabel custom')
-        expect(line?.getAttribute('class')).toContain('lineLabel custom')
+        const path = result.querySelector('path')
+        expect(text?.getAttribute('class')).toContain('label bracketLabel custom')
+        expect(path?.getAttribute('class')).toContain('bracketLabel custom')
     })
 })